Glade files are not compat from 2->3
Even using Gtk.Builder, the version of glade
that supports AnacondaWidgets does not support
gtk2 at all and forces a
'requires lib=Gtk+ verison=3.0'. That doesn't
work on gtk2 of course, but it also means that
glade flags things like 'VBox' as deprecated
in lieu of 'Box'.

That likely means gtk2/gtk3 compat with the ui
builder files is not possible.

This commit starts moving this branch towards gtk3
only.